Large scale reorganization of network onboarding app
