+++ network-onboarding/network-onboarding-server/config/.gitignore
@@ -1,0 +1,1 @@
+testdeploy.*.yaml
