+++ network-onboarding/network-onboarding-server/dapp/application/contracts/Application.sol
@@ -12,14 +12,23 @@ contract Application {
   
   string public name;                       // application name
   address public ownerOrganization;         // address of the org that owns this app 
-  mapping(string => address) organizations;   // other orgs that participate in this app
+  mapping(string => address) public organizations; // other orgs that participate in this app; NOT QUERYABLE
 
   constructor(
     string _name,
-    string _ownerOrganization
+    address _ownerOrganization
   ) {
     owner = msg.sender;
     name = _name;
     ownerOrganization = _ownerOrganization;
+  }
+
+  function addOrganization(string _name, address _address) {
+    if (organizations[_name] != address(0))
+      organizations[_name] = _address;
+  }
+
+  function getApplicationOwner() returns (address) {
+    return ownerOrganization;
   }
 }
