+++ network-onboarding/network-onboarding-server/dapp/dapp/contracts/NetworkOnboardingDapp.sol
@@ -1,7 +1,7 @@
-pragma solidvm 3.0;
+// pragma solidvm 3.0;
 
 import "/dapp/permission/contracts/NetworkOnboardingPermissionManager.sol";
-import "/dapp/user/contracts/NetworkOnboardingUserManager.sol";
+import "/dapp/user-manager/contracts/NetworkOnboardingUserManager.sol";
 import "/dapp/organization/contracts/OrganizationManager.sol";
 import "/dapp/organization/membership/contracts/OrganizationMembershipManager.sol";
 
@@ -26,11 +26,8 @@ contract NetworkOnboardingDapp {
     owner = msg.sender;
     permissionManager = new NetworkOnboardingPermissionManager(msg.sender, msg.sender);
     userManager = new NetworkOnboardingUserManager(permissionManager);
-    organizationManager = new OrganizationManager(permissionManager); // userManager?
+    organizationManager = new OrganizationManager(permissionManager, userManager);
     organizationMembershipManager = new OrganizationMembershipManager(permissionManager, organizationManager); // userManager?
-     
-    applicationManager = new ApplicationManager(permissionManager, /* TODO */ address(0)); 
-
-    // TODO: add the rest of the manager initializations
+    applicationManager = new ApplicationManager(permissionManager, organizationManager);
   }
 }
