+++ network-onboarding/network-onboarding-server/dapp/organization/contracts/Organization.sol
@@ -1,5 +1,3 @@
-import "/dapp/user/contracts/NetworkOnboardingUser.sol";
-
 /**
  * Organization data contract
  * The Organization contract stores data pertinent to the organization, it will also expose a number
@@ -9,23 +7,16 @@ import "/dapp/user/contracts/NetworkOnboardingUser.sol";
  * Holds data for an organization, including the current members of the organization
  */
 contract Organization {
-    address public owner;    // Ties the Organization contract with the OrganizationManager
-    // string public commonName;
-    string public certificateString;    // Intermediate certificate for an organization
-    address[] public members;
+    address public owner;       // The creator of the contract, i.e. OrganizationManager
+    // IMPORTANT!!! There is an implicit state that this contract has a X.509 certificate registed to its address
+    address[] public members;   // TODO Use a better data structure (maybe a mapping)
 
-    constructor(string _certificateString) {
+    constructor() {
         owner = msg.sender;
-        // commonName = parseCert(_commonName)["commonName"];
-        certificateString = _certificateString;
         members = [];
     }
 
-    function addMember(address _user) returns (uint256) {
-        // user exists
-        NetworkOnboardingUser user = NetworkOnboardingUser(_member);
-        if (address(user) == 0) return (RestStatus.NOT_FOUND, 0);
-
+    function addMember(address _userAddress) returns (uint256) {
         // TODO Add more verification to adding members
 
         members.push(_member);
