+++ network-onboarding/network-onboarding-server/dapp/user-manager/test/networkOnboardingUserManager.test.js
@@ -1,0 +1,135 @@
+import { util, rest, assert } from '/blockapps-rest-plus'
+import RestStatus from 'http-status-codes'
+import config from '/load.config'
+import constants from '/helpers/constants'
+import { getCurrentEnode } from '/helpers/enodeHelper'
+import { getRoles } from '/helpers/enums'
+import oauthHelper from '/helpers/oauthHelper'
+import dotenv from 'dotenv'
+
+import networkOnboardingPermissionManagerJs from '/dapp/permission/permissionManager'
+import networkOnboardingUserManagerJs from '/dapp/user-manager/networkOnboardingUserManager'
+
+const options = { config }
+const roles = getRoles()
+
+const loadEnv = dotenv.config()
+assert.isUndefined(loadEnv.error)
+
+describe('NetworkOnboarding User Manager', function () {
+  this.timeout(config.timeout)
+
+  let enodeAddress
+  let networkAdmin
+  let orgAdmin
+  let permissionManagerContract
+
+  before(async () => {
+    let networkAdminToken
+    try {
+      networkAdminToken = await oauthHelper.getUserToken(`${process.env.NETWORK_ADMIN_NAME}`, `${process.env.TEST_USER_PASSWORD}`)
+    } catch (e) {
+      console.error('ERROR: Unable to fetch the admin token, check your OAuth settings in config', e)
+      throw e
+    }
+    const networkAdminCredentials = { token: networkAdminToken }
+    networkAdmin = await rest.createUser(networkAdminCredentials, options)
+    const args = {
+      admin: networkAdmin.address,
+      master: networkAdmin.address,
+    }
+    permissionManagerContract = await networkOnboardingPermissionManagerJs.uploadContract(networkAdmin, args, options)
+    // grant network admin role
+    await permissionManagerContract.grantNetworkAdminRole({ user: networkAdmin })
+    enodeAddress = getCurrentEnode()
+
+    let orgAdminToken
+    try {
+      orgAdminToken = await oauthHelper.getUserToken(`${process.env.ORGANIZATION_ADMIN_NAME}`, `${process.env.TEST_USER_PASSWORD}`)
+    } catch (e) {
+      console.error('ERROR: Unable to fetch the org admin token, check your OAuth settings in config', e)
+      throw e
+    }
+    const orgAdminCredentials = { token: orgAdminToken }
+    orgAdmin = await rest.createUser(orgAdminCredentials, options)
+    await permissionManagerContract.grantOrganizationAdminRole({ user: orgAdmin })
+
+  })
+
+  it('Create NetworkOnboarding User Manager', async () => {
+    const contract = await networkOnboardingUserManagerJs.uploadContract(networkAdmin, {
+      permissionManager: permissionManagerContract.address,
+      enodeAddress,
+    }, options)
+    assert.notEqual(contract.address, constants.zeroAddress, 'Contract address must be not zero')
+
+    const {
+      permissionManager,
+    } = await contract.getState()
+
+    assert.equal(permissionManager, permissionManagerContract.address, 'permissionManager')
+  })
+
+  describe('NetworkOnboarding User Create/Update', () => {
+    let userArgs
+    let contract
+    let orgAdminContract
+
+
+    before(async () => {
+      contract = await networkOnboardingUserManagerJs.uploadContract(networkAdmin, {
+        permissionManager: permissionManagerContract.address,
+        enodeAddress,
+      }, options)
+
+      // register network admin user
+      await contract.registerUser({ 
+        userAddress: networkAdmin.address, 
+        userCertificate: '-----BEGIN CERTIFICATE-----\nMIIBiDCCAS2gAwIBAgIQCgO76hC29iXEFXJNco5ekjAMBggqhkjOPQQDAgUAMEYx\nDDAKBgNVBAMMA2RhbjEMMAoGA1UEBgwDVVNBMRIwEAYDVQQKDAlibG9ja2FwcHMx\nFDASBgNVBAsMC2VuZ2luZWVyaW5nMB4XDTIxMDMxODE1NDgwN1oXDTIyMDMxODE1\nNDgwN1owRjEMMAoGA1UEAwwDZGFuMQwwCgYDVQQGDANVU0ExEjAQBgNVBAoMCWJs\nb2NrYXBwczEUMBIGA1UECwwLZW5naW5lZXJpbmcwVjAQBgcqhkjOPQIBBgUrgQQA\nCgNCAAQY4p67l1IIEUdVC7L+rUDwF5Nv30bze0NV5y8ced7qwp+YFk3UAiOGkcYo\n7ba8F92rd0yf9AGpvZN1H3Dda8xdMAwGCCqGSM49BAMCBQADRwAwRAIgbKXO8tZ5\noPhBusPQFkNEQDnLO/MRru4KjtCpPnVb5sACIE0TwBJ7yeIGuPc/8G50/858Pf3a\n0t1hHbhYnJarPkNA\n-----END CERTIFICATE-----',
+        role: roles.NETWORK_ADMIN
+      })
+
+      // register the org admin user
+      await contract.registerUser({
+        userAddress: orgAdmin.address,
+        userCertificate: '-----BEGIN CERTIFICATE-----\nMIIBiDCCAS2gAwIBAgIQCgO76hC29iXEFXJNco5ekjAMBggqhkjOPQQDAgUAMEYx\nDDAKBgNVBAMMA2RhbjEMMAoGA1UEBgwDVVNBMRIwEAYDVQQKDAlibG9ja2FwcHMx\nFDASBgNVBAsMC2VuZ2luZWVyaW5nMB4XDTIxMDMxODE1NDgwN1oXDTIyMDMxODE1\nNDgwN1owRjEMMAoGA1UEAwwDZGFuMQwwCgYDVQQGDANVU0ExEjAQBgNVBAoMCWJs\nb2NrYXBwczEUMBIGA1UECwwLZW5naW5lZXJpbmcwVjAQBgcqhkjOPQIBBgUrgQQA\nCgNCAAQY4p67l1IIEUdVC7L+rUDwF5Nv30bze0NV5y8ced7qwp+YFk3UAiOGkcYo\n7ba8F92rd0yf9AGpvZN1H3Dda8xdMAwGCCqGSM49BAMCBQADRwAwRAIgbKXO8tZ5\noPhBusPQFkNEQDnLO/MRru4KjtCpPnVb5sACIE0TwBJ7yeIGuPc/8G50/858Pf3a\n0t1hHbhYnJarPkNA\n-----END CERTIFICATE-----',
+        role: roles.ORG_ADMIN,
+      })
+
+      // bind the manager contract to org admin user, for testing permission checks
+      orgAdminContract = networkOnboardingUserManagerJs.bindAddress(orgAdmin, contract.address, options)
+    })
+
+    it('Create NetworkOnboarding User - 201 - CREATED', async () => {
+      // create NetworkOnboarding User
+      const networkOnboardingUser = await contract.registerUser({
+        userAddress: '1234',
+        userCertificate: '-----BEGIN CERTIFICATE-----\nMIIBiDCCAS2gAwIBAgIQCgO76hC29iXEFXJNco5ekjAMBggqhkjOPQQDAgUAMEYx\nDDAKBgNVBAMMA2RhbjEMMAoGA1UEBgwDVVNBMRIwEAYDVQQKDAlibG9ja2FwcHMx\nFDASBgNVBAsMC2VuZ2luZWVyaW5nMB4XDTIxMDMxODE1NDgwN1oXDTIyMDMxODE1\nNDgwN1owRjEMMAoGA1UEAwwDZGFuMQwwCgYDVQQGDANVU0ExEjAQBgNVBAoMCWJs\nb2NrYXBwczEUMBIGA1UECwwLZW5naW5lZXJpbmcwVjAQBgcqhkjOPQIBBgUrgQQA\nCgNCAAQY4p67l1IIEUdVC7L+rUDwF5Nv30bze0NV5y8ced7qwp+YFk3UAiOGkcYo\n7ba8F92rd0yf9AGpvZN1H3Dda8xdMAwGCCqGSM49BAMCBQADRwAwRAIgbKXO8tZ5\noPhBusPQFkNEQDnLO/MRru4KjtCpPnVb5sACIE0TwBJ7yeIGuPc/8G50/858Pf3a\n0t1hHbhYnJarPkNA\n-----END CERTIFICATE-----',
+        role: roles.ORG_ADMIN,
+      }, options)
+      assert.equal(networkOnboardingUser, RestStatus.CREATED, 'restStatus')
+    })
+
+    it('Create NetworkOnboarding User - 409 - CONFLICT', async () => {
+      // create NetworkOnboarding User with the same parameters
+      await assert.restStatus(async () => {
+        await contract.registerUser({
+          userAddress: '1234',
+          userCertificate: '-----BEGIN CERTIFICATE-----\nMIIBiDCCAS2gAwIBAgIQCgO76hC29iXEFXJNco5ekjAMBggqhkjOPQQDAgUAMEYx\nDDAKBgNVBAMMA2RhbjEMMAoGA1UEBgwDVVNBMRIwEAYDVQQKDAlibG9ja2FwcHMx\nFDASBgNVBAsMC2VuZ2luZWVyaW5nMB4XDTIxMDMxODE1NDgwN1oXDTIyMDMxODE1\nNDgwN1owRjEMMAoGA1UEAwwDZGFuMQwwCgYDVQQGDANVU0ExEjAQBgNVBAoMCWJs\nb2NrYXBwczEUMBIGA1UECwwLZW5naW5lZXJpbmcwVjAQBgcqhkjOPQIBBgUrgQQA\nCgNCAAQY4p67l1IIEUdVC7L+rUDwF5Nv30bze0NV5y8ced7qwp+YFk3UAiOGkcYo\n7ba8F92rd0yf9AGpvZN1H3Dda8xdMAwGCCqGSM49BAMCBQADRwAwRAIgbKXO8tZ5\noPhBusPQFkNEQDnLO/MRru4KjtCpPnVb5sACIE0TwBJ7yeIGuPc/8G50/858Pf3a\n0t1hHbhYnJarPkNA\n-----END CERTIFICATE-----',
+          role: roles.ORG_ADMIN,
+        }, options)
+      }, RestStatus.CONFLICT)
+    })
+    
+    it('Create NetworkOnboarding User - 403 - FORBIDDEN - org admin creating network admin', async () => {
+      // orgAdmin tries to create a network admin
+      await assert.restStatus(async () => {
+        await orgAdminContract.registerUser({
+          userAddress: '1236',
+          userCertificate: '-----BEGIN CERTIFICATE-----\nMIIBiDCCAS2gAwIBAgIQCgO76hC29iXEFXJNco5ekjAMBggqhkjOPQQDAgUAMEYx\nDDAKBgNVBAMMA2RhbjEMMAoGA1UEBgwDVVNBMRIwEAYDVQQKDAlibG9ja2FwcHMx\nFDASBgNVBAsMC2VuZ2luZWVyaW5nMB4XDTIxMDMxODE1NDgwN1oXDTIyMDMxODE1\nNDgwN1owRjEMMAoGA1UEAwwDZGFuMQwwCgYDVQQGDANVU0ExEjAQBgNVBAoMCWJs\nb2NrYXBwczEUMBIGA1UECwwLZW5naW5lZXJpbmcwVjAQBgcqhkjOPQIBBgUrgQQA\nCgNCAAQY4p67l1IIEUdVC7L+rUDwF5Nv30bze0NV5y8ced7qwp+YFk3UAiOGkcYo\n7ba8F92rd0yf9AGpvZN1H3Dda8xdMAwGCCqGSM49BAMCBQADRwAwRAIgbKXO8tZ5\noPhBusPQFkNEQDnLO/MRru4KjtCpPnVb5sACIE0TwBJ7yeIGuPc/8G50/858Pf3a\n0t1hHbhYnJarPkNA\n-----END CERTIFICATE-----',
+          role: roles.NETWORK_ADMIN,
+        }, options)
+      }, RestStatus.FORBIDDEN)
+    })
+  })
+})
