+++ network-onboarding/network-onboarding-server/package.json
@@ -44,11 +44,11 @@
     "test:permission": "mocha --require @babel/register dapp/permission/test -b",
     "test:organization": "mocha --require @babel/register dapp/organization/test -b",
     "test:organization-membership": "mocha --require @babel/register dapp/organization/membership/test -b",
-    "test:user": "mocha --require @babel/register dapp/user/test -b",
+    "test:user-manager": "mocha --require @babel/register dapp/user-manager/test -b",
     "test:application": "mocha --require @babel/register dapp/application/test -b",
     "test:dapp": "mocha --require @babel/register dapp/dapp/test/dapp.test.js -b",
     "test:e2e": "mocha --require @babel/register dapp/dapp/test/e2e.test.js -b",
-    "test": "yarn test:permission && yarn test:organization && yarn test:organization-membership && yarn test:application && yarn test:user && yarn test:dapp && yarn test:e2e",
+    "test": "yarn test:permission && yarn test:organization && yarn test:application && yarn test:user-manager && yarn test:dapp && yarn test:e2e",
     "clean": "rm -f config/*.deploy.yaml config/testdeploy.*"
   }
 }
